chore: add analytics mixin
