+++ src/config.ts
@@ -1,0 +1,5 @@
+//TODO reformat
+export const config = {
+    pollingInterval : 1000,
+    LOAN_ID_IDX: 2
+}
